pyproject.toml
[project]
name = "xml-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.14.0",
    "openai>=1.107.3",
    "python-dotenv>=1.1.1",
]
